refactor: TalentTest 중복되는 코드를 활용하여 메소드 생성
패키지 이름 변경
api -> skhuedin
entity -> domain

com.skhuedin.api
 -> com.skhuedin.skhuedin

com.skhuedin.api.entity
 -> com.skhuedin.skhuedin.domain

패키지 이름은 후에 스프링 모듈화를 고려하여 좀 더 고민을 보류
