+++ src/main/java/com/skhuedin/skhuedin/SkhuedInApplication.java
@@ -1,4 +1,4 @@
-package com.skhuedin.api;
+package com.skhuedin.skhuedin;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
