+++ src/main/java/com/skhuedin/skhuedin/domain/Roadmap.java
@@ -1,4 +1,4 @@
-package com.skhuedin.api.entity;
+package com.skhuedin.skhuedin.domain;
 
 import lombok.AccessLevel;
 import lombok.Builder;
@@ -17,20 +17,22 @@ import javax.persistence.ManyToOne;
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Follow extends BaseEntity {
+public class Roadmap extends BaseEntity {
 
     @Id
     @GeneratedValue
-    @Column(name = "follow_id")
+    @Column(name = "roadmap_id")
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id")
-    private User user;
+    @JoinColumn(name = "talent_id")
+    private Talent talent;
 
     @Builder
-    public Follow(User user) {
-        Assert.hasText(String.valueOf(user), "user 값은 필수입니다. ");
-        this.user = user;
+    public Roadmap(Talent talent) {
+
+        Assert.notNull(talent, "talent 필수입니다. ");
+
+        this.talent = talent;
     }
 }
